backend/requirements.txt
alembic==1.8.0
astroid==2.2.5
Authlib==1.0.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
ecdsa==0.13.2
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
future==0.17.1
greenlet==1.1.2
idna==3.3
isort==4.3.18
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.4.0
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.3.1
pylint==2.3.1
python-dotenv==0.20.0
python-jose==3.3.0
python-jose-cryptodome==1.3.2
requests==2.27.1
rsa==4.8
six==1.12.0
SQLAlchemy==1.4.37
typed-ast==1.5.4
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.11.1
